requirements.txt
pandas==2.2.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
streamlit==1.41.1
plotly==5.24.1
python-dotenv==1.0.1
pytest==8.3.4
numpy==2.2.1
scikit-learn==1.6.1
